Finalizacion de boletin 10 e inicio do 11.

+++ FRMCode/NetBeansProjects/Boletin 10/Boletin10_1/src/boletin10_1/Boletin10_1.java
@@ -14,7 +14,7 @@ public class Boletin10_1 {
             }while(n<0 &&n>50);
             
                 do{
-                    i=(int) Float.parseFloat(JOptionPane.showInputDialog("Xogador 1, introduza o numero de intentos que terá o xogador 2:"));
+                    i=Integer.parseInt(JOptionPane.showInputDialog("Xogador 1, introduza o numero de intentos que terá o xogador 2:"));
                 }while(i<1 &&i>49);
                 Xogador x1=new Xogador(n,i);
                 x1.xogar();
